requirements.txt
streamlit==1.40.0
huggingface_hub>=0.20.0
sentence-transformers>=2.2.0
openai>=1.0.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
pypdf>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
